优化客户新增处理，商机模板新增逻辑处理

+++ EMT.DoneNOW.Core/SYS/sys_form_tmpl_opportunity.cs
@@ -15,7 +15,7 @@ namespace EMT.DoneNOW.Core
 
         [Key]
         [DataMember]
-        public Int32 id { get; set; }
+        public Int64 id { get; set; }
         [DataMember]
         public Int32 form_tmpl_id { get; set; }
         [DataMember]
